Completed Frame UI Creation and Functionality
Creates a Java GUI frame that can be closed, iconified, scaled, moved, and added to. Should the user press the X button on the top right it shall either exit the application or prompt for the user to save, should saving be an option.  Currently there is no saving, but the ability was added.  I also added a .jar file for testing purposes, and javadocs for documentation.