Added the Deletion feature
Updated documentation files and documented the code.  Also updated the Readme to be more up-to-date.  Spread apart the features so they could be worked on easier later.  The deletion feature is functional, however requires an item to be selected, so the selection feature must be complete to thoroughly test.

Whilst tampering with the buttons I realized that should I hold the insert button, and the object continuously gets created, then the application will crash.  I believe that some limitation should be set either for the speed of creation OR for the amount of ObjectUIs created.  Or a different approach is used, such as instead of creating one at a time, we can create groups at a time.  This concept can be worked on later, as the creation feature has yet to be completed.