Integrated Seletion and Edit View
ObjectUI was reverted, as there isn't a need for two references to whether it is selected.
TEKActionAdapter was adjusted to include both Edit and Select All.
TEKEditView was changed to a .java file.
TEKFrame was altered to add an imageIcon and an updated WindowListener.  The sample objects were removed, and the popupMenu can now be obtained from this class.
TEKFrameAdapter replaces TEKWindowAdapter since TEKFrame may require more than just a WindowListener later.
TEKLabel was altered to remove the getObjectFromText method and move setOpaque to within the initializer method.
TEKLabelAdapter was altered to activate the popup menu upon right click and selection on left click.
TEKManagement was altered to fill a sample creation scheme, which will be replaced later, and add selection and edit capabilities.
TEKMenuBar was altered to become a builder for JMenus and JMenuItems, although it can still be used for a menuBar.  The same occurred for the TEKToolBar.
TEKPanel was altered to have more built-in functionality, such as the ability to getObjectFromLabel or add multiple objects at once without calling the addObject method multiple times.  The displayObjects method was removed to incorporate a generation scheme that is similar but only fires when necessary.
TEKPanelAdapter was created to allow for the TEKPanel to perform desired actions, such as PopupMenusactivation.
TEKPopupMenu was created to support the usage of PopupMenus for creating TEKEditView with the selected TEKLabel.
TEK-GUI.jar was updated to be a sample representation of what has been done throughout the sprint so far, it may be updated later as well.

Tests:
For tests, the Helper class was updated to support customizability so that developers or users may have a better experience.  A rough emergency procedure was introduced utilizing JDialogs, however it requires touchScreen and will be removed in a later interation in favor of encapsulating the TEKFrame in another JFrame by forcing it to be a JInternalFrame.
The Internal class showcases the ability to force a JFrame into a JInternalFrame.
SampleTestingInterface shows the usage of Internal and TEKFrame to create a safer environment for the developers whilst testing.  IN-PROGRESS
The MainTester class will be used for future testing, but will require FrameFunctionalityTest.java to be setup to combine all tests and support the customizable features of Helper.  More classes are expected to be created for testing, with each being connected to MainTester.