Create ImprovedSampleTestingInterface.jar
Added an improved jar file that has more functionality, such as taking in System.out.println(...) and having a button to exit.  I have yet to integrate the other features for TEKGUI, and this will not work correctly when another JFrame is opened.  JFrame windowListeners will also not work within the JInternalFrames.

I have not polished the code, if you want to see it, you can by extracting the jar, but it is not complete.