Moved everything into a package
Updated the TEKGUI jar and TestSuite jar, hopefully I have patched the PopupMenu issue, if it arises again, please let me know.

The TestSuite is not complete, but it is getting there.  Added a Test abstract class that will later be linked to TestFrame in some way.

The only thing major about this commit would be organization.  Hopefully this will make it easier.