Populated Popup Menu w/ Common Features
Removed creation from MenuBar and ToolBar to allow for more customizability.
Updated the classes for TestSuite to have more features, Testing is still incomplete.  It will use the TestTree class which has yet to be completed.
TEKPopupMenu is now ready for Common Features to be added, although I will still need to wait for their completion prior to finishing the feature completely, as it requires setEnable to be responsive to the user's current situation.
TEKFile was updated to fix an issue that arose when using the TestSuite, Saving or Opening a file caused the developer to be stuck in fullScreen should they ALT-TAB out prior to inputting or exiting the fileChooser.
TEKLabelAdapter and TEKPanelAdapter were changed to make popupMenu more responsive, although there are still issues as TEKFile.getFrame() may return null.