Connected HTML extractor to Files
Added a good amount of files that won't be touched until next semester, most likely, like Layouts, and FolderUI, and ui effects (Blur, Validate, Loading).  Ignore these files for now.

Located in Adapter, although that aren't adapters, they will be used by them.
Added a practically empty UndoManger and UndoableAction to kickstart that feature.
Added a ShortcutSystem file with a slight template, not fleshed out yet, for a future feature.
Adjusted TEKLabel and TEKPanel to incorporate the label inside ObjectUI.  Will probably need to do more later.

Most importantly, adjusted TEKFile to have more options when interacting with files, as the user should not always be required to input a file or output to a specific file.  Although the methods are yet to be completed.  ObjectUI content can be obtained and generated from a file, once initFile is filled in, the currentTime was modified to be the last time the file was modified.

HTML Extractor and Comment Cleaner were adjusted to work with Files or be slightly easier to use.