Added Undo/Redo Base
Adds types of actions, such as Creation, Deletion, Move, Adjust, and Unknown and attaches it to an object.  The object could be a list.

Undo/Redo can stack to 50 per, so 100 total.  The current approach does not clear redo when another action is placed onto undo.  canRedo and canUndo can now be used to determine whether various UI elements are unable to be selected, such as ToolBar and PopupMenu.