Added a finder that doesn't function yet
I did a few tweaks to integrate some of the features.
For recent files, I had to remove some of the stuff within TEKFrame and alter TEKMenuBar a little, adjusting TEKFile to ensure that when a file is opened, it is remembered.
For the highlighting feature, I added upon it to allow for the keyBoard to be used to navigate (TAB).
For the rectangular select feature, I added LayerUI to make it paint more smoothly, it was painting under and when I tried to make it paint on top it was inconsistent, thus that was the solution.  Which may have broken some things sadly.

For ShortcutSystem I fixed an error I didn't spot when I was reviewing the code, that the method addObjects did not exist and it was called addObject.  I am calling the trigger method in TEKFinder to open up the view.
For TEKFinder, I added buttons for navigating and entering in, such as Find and Replace for inputs, Prev, Next, Expand, and Exit for buttons.  Find, Replace, Next, and Prev do not work properly as of this commit, so I have commented some of the code.
For TEKLabelAdapter, I added a focus and keyListener so the keyboard could traverse through and interact.
For TEKPanelAdapter I basically added something to allow users to smoothly transition from keyboard to mouse, although it probably needs to be adjusted slightly.
For ButtonBuilder, I had to learn the exact dimensions and color of the default border around a button, although the color isn't perfect, and use them in tandem with a focusListener to ensure that users can see which button is active.  I probably need to adjust it though.  I added a keyListener that checks for the enter key when the button is focused, which allows for users to use purely keyboard.  I added a skinButton method, which acts exactly like FleshButton, but I just moved it to a better place.
For TEKLabel, I just made it focusable and added the listeners.
For TEKToolBar did a little adjustment to make it look neater.

The recentFiles.dat file is recommended to be located with the TEKGUI-3.jar within a folder.