Updated TEKPopupMenu
TEKPopupMenu can now take in multiple objectUIs at once, also made the EditView able to take in multiple.  It still needs further work.

TEKPanel may get updated to be like a grid to make it more efficient.  Update row/column each time an object is generated, and reset to 0/0 when removing an item.

TEKLabelAdapter was adjusted since it didn't allow for multiple objects to be selected individually.